gost-chat/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic-settings==2.7.1
jinja2==3.1.5
PyMuPDF>=1.25.0,<2.0.0
qdrant-client>=1.9,<2
sentence-transformers>=3.0,<4
torch>=2.3,<3
